Fix https://github.com/gnustavo/JIRA-Client/issues/1
The logic implemented in v0.38 to allow for noon-standard WSDL
descriptors was flawed. It broke invokations where the JIRA baseURL
had a path part. Now I went back to the original semantics and
implemented an alternative invokation style based on named arguments
via a hash-ref. It is more flexible and extensible.