Implement the method filter_issues_unsorted
The filter_issues method sorts the list of issues by key before
returning it. This was a bad idea because the user may want the list
sorted in another way and even specify this via JQL, but the method
would break the sorting.

Bad as it is I'm reluctant to change it because it could break some code
already depending on this sorting.

So, this implements a new method called filter_issues_unsorted which
simply returns the list of issues as is. The filter_issues method is now
reimplemented on top of the new one.

Kudos to Alan Quille for bringing this issue to my attention.